Move get_proc_address to renderer file
