+++ src/renderers/mod.rs
@@ -3,3 +3,4 @@
 pub mod imgui_dx11;
 pub mod imgui_dx12;
 pub mod imgui_dx9;
+pub mod imgui_opengl3;
